feat(adapters): M15.2 PostgreSQL adapter — PgStorage with all 16 port traits (#120)
* feat(adapters): M15.2 — add PgStorage PostgreSQL adapter

- Add `postgres` feature to diesel workspace dependency
- Create crates/gyre-adapters/src/postgres/ with PgStorage implementing
  all 16 port traits (project, agent, repo, task, merge_request,
  merge_queue, review, activity, analytics/cost, audit, agent_commit,
  worktree, user/api_key, network_peer)
- Share migrations/ directory between SQLite and PostgreSQL backends;
  fix DOUBLE → DOUBLE PRECISION for PostgreSQL DDL compatibility
- Wire GYRE_DATABASE_URL=postgres://... detection into build_state():
  postgres:// prefix → PgStorage, sqlite:// prefix → SqliteStorage,
  unset → in-memory


* fix(server): include git_refs.rs in M15.2 branch

Adds the missing crates/gyre-server/src/git_refs.rs (M13.6 custom
git refs module) so the branch compiles cleanly after rebase on main.


* fix(server): remove git_refs.rs from M15.2 branch — let PR #121 handle it

Avoids dead_code clippy warnings on this branch since the callers
(spawn/complete) live in PR #121's branch, not here.


* style: cargo fmt — remove blank lines from lib.rs


---------

+++ crates/gyre-adapters/migrations/2024-01-01-000001_initial_schema/up.sql
@@ -157,7 +157,7 @@ CREATE TABLE IF NOT EXISTS cost_entries (
     agent_id TEXT NOT NULL,
     task_id TEXT,
     cost_type TEXT NOT NULL,
-    amount DOUBLE NOT NULL,
+    amount DOUBLE PRECISION NOT NULL,
     currency TEXT NOT NULL,
     timestamp BIGINT NOT NULL
 );
